Only use the LeaderBoard that displays Points
