+++ package.json
@@ -45,26 +45,26 @@
     "node": "8.11.4"
   },
   "dependencies": {
-    "aws-sdk": "^2.357.0",
+    "aws-sdk": "^2.377.0",
     "axios": "^0.18.0",
     "bootstrap": "^4.1.3",
-    "express": "^4.16.3",
+    "express": "^4.16.4",
     "lodash": "^4.17.11",
-    "npm": "^6.4.1",
+    "npm": "^6.5.0",
     "npm-run-all": "^4.1.5",
     "randomstring": "^1.1.5",
-    "react": "^16.6.0",
-    "react-dom": "^16.6.0",
-    "react-redux": "^5.1.0",
+    "react": "^16.6.3",
+    "react-dom": "^16.6.3",
+    "react-redux": "^5.1.1",
     "react-router": "^3.2.1",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "^2.0.5",
+    "react-scripts": "^2.1.1",
     "reactstrap": "^6.5.0",
     "redux": "^4.0.1",
     "redux-form": "^7.4.2",
     "redux-thunk": "^2.3.0",
-    "socket.io": "^2.1.1",
-    "socket.io-client": "^2.1.1"
+    "socket.io": "^2.2.0",
+    "socket.io-client": "^2.2.0"
   },
   "browserslist": [
     ">0.2%",
@@ -73,6 +73,6 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "redux-devtools-extension": "^2.13.5"
+    "redux-devtools-extension": "^2.13.7"
   }
 }
